Fix deb package generation and example binary
This commit fixes a conceptual error on package generation: the
package should not only install the .so library, but also the header
files necessary for acessing the methods declared in the library.
With this, there was also another conceptual mistake on the binary
that served as an example of the library: the inclusion should not
be as a "local" lib, but as a "system" lib, using <header.h> instead
of "header.h" in the main file.